# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: PawZone API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /listings:
    get:
      operationId: getListings
      tags: [listings]
      summary: Get all approved listings
      parameters:
        - name: category
          in: query
          schema:
            type: string
        - name: breed
          in: query
          schema:
            type: string
        - name: minPrice
          in: query
          schema:
            type: number
        - name: maxPrice
          in: query
          schema:
            type: number
        - name: search
          in: query
          schema:
            type: string
        - name: sellerId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of listings
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Listing"
    post:
      operationId: createListing
      tags: [listings]
      summary: Create a new pet listing (seller only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateListingBody"
      responses:
        "201":
          description: Listing created (pending admin approval)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Listing"
  /listings/{id}:
    get:
      operationId: getListing
      tags: [listings]
      summary: Get a listing by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Listing details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListingDetail"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      operationId: updateListing
      tags: [listings]
      summary: Update a listing (seller only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateListingBody"
      responses:
        "200":
          description: Listing updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Listing"
    delete:
      operationId: deleteListing
      tags: [listings]
      summary: Delete a listing (seller only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /cart:
    get:
      operationId: getCart
      tags: [cart]
      summary: Get current user's cart
      responses:
        "200":
          description: Cart items
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cart"
    post:
      operationId: addToCart
      tags: [cart]
      summary: Add item to cart
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddToCartBody"
      responses:
        "200":
          description: Cart updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cart"
  /cart/{listingId}:
    put:
      operationId: updateCartItem
      tags: [cart]
      summary: Update cart item quantity
      parameters:
        - name: listingId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateCartItemBody"
      responses:
        "200":
          description: Cart updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cart"
    delete:
      operationId: removeFromCart
      tags: [cart]
      summary: Remove item from cart
      parameters:
        - name: listingId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Cart updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cart"
...
  /notifications/{id}/read:
    post:
      operationId: markNotificationRead
      tags: [dashboard]
      summary: Mark notification as read
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Notification marked read
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required:
        - error
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required:
        - success
...
    User:
      type: object
      properties:
        id:
          type: integer
        role:
          type: string
          enum: [buyer, seller, transporter, admin]
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        address:
          type: string
        city:
          type: string
        state:
          type: string
        pincode:
          type: string
        country:
          type: string
        latitude:
          type: number
        longitude:
          type: number
        status:
          type: string
          enum: [pending, approved, blocked]
        sellerId:
          type: string
          nullable: true
        sellerScore:
          type: number
          nullable: true
        profilePhoto:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - role
        - name
        - email
        - status
        - createdAt
    Listing:
      type: object
      properties:
        id:
          type: integer
        sellerId:
          type: integer
        sellerName:
          type: string
        category:
          type: string
          enum: [dogs, cats, birds, fish]
        breed:
          type: string
        price:
          type: number
        quantity:
          type: integer
        availableQuantity:
          type: integer
        vaccinated:
          type: boolean
        vaccinationDetails:
          type: string
          nullable: true
        photos:
          type: array
          items:
            type: string
        videoUrl:
          type: string
          nullable: true
        description:
          type: string
        status:
          type: string
          enum: [pending, approved, rejected, sold_out]
        petCode:
          type: string
          nullable: true
        address:
          type: string
        city:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - sellerId
        - sellerName
        - category
        - breed
        - price
        - quantity
        - availableQuantity
        - vaccinated
        - photos
        - description
        - status
        - address
        - city
        - createdAt
    ListingDetail:
      allOf:
        - $ref: "#/components/schemas/Listing"
        - type: object
          properties:
            sellerInfo:
              $ref: "#/components/schemas/User"
            reviews:
              type: array
              items:
                $ref: "#/components/schemas/Review"
            avgRating:
              type: number
    CreateListingBody:
      type: object
      properties:
        category:
          type: string
          enum: [dogs, cats, birds, fish]
        breed:
          type: string
        price:
          type: number
        quantity:
          type: integer
        vaccinated:
          type: boolean
        vaccinationDetails:
          type: string
        photos:
          type: array
          items:
            type: string
        videoUrl:
          type: string
        description:
          type: string
        address:
          type: string
        city:
          type: string
      required:
        - category
        - breed
        - price
        - quantity
        - vaccinated
        - description
        - address
        - city
    CartItem:
      type: object
      properties:
        listingId:
          type: integer
        listing:
          $ref: "#/components/schemas/Listing"
        quantity:
          type: integer
        subtotal:
          type: number
        platformFee:
          type: number
      required:
        - listingId
        - listing
        - quantity
        - subtotal
        - platformFee
    Cart:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/CartItem"
        subtotal:
          type: number
        platformFee:
          type: number
        total:
          type: number
        itemCount:
          type: integer
      required:
        - items
        - subtotal
        - platformFee
        - total
        - itemCount
    AddToCartBody:
      type: object
      properties:
        listingId:
          type: integer
        quantity:
          type: integer
      required:
        - listingId
        - quantity
    UpdateCartItemBody:
      type: object
      properties:
        quantity:
          type: integer
      required:
        - quantity
...
    Review:
      type: object
      properties:
        id:
          type: integer
        orderId:
          type: integer
        buyerId:
          type: integer
        buyerName:
          type: string
        sellerId:
          type: integer
        rating:
          type: integer
        comment:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - orderId
        - buyerId
        - buyerName
        - sellerId
        - rating
        - createdAt
...